fix minor issues and typos

+++ 25-26-2-pr1-PR1Statement/src/main/java/edu/uoc/ds/adt/util/ArchimedeanSpiral.java
@@ -29,7 +29,7 @@ public class ArchimedeanSpiral extends JPanel {
         for (int i = 0; i <= currentIndex && i < points.length; i++) {
             double x = points[i].x();
             double y = points[i].y();
-            g2d.fillOval((int) (x - size / 2), (int) (y - size / 2), size, size);
+            g2d.fillOval((int) (x - (double) size / 2), (int) (y - (double) size / 2), size, size);
         }
     }
